[#4] Create authentication flow (#11)

+++ backend/src/main/kotlin/com/demo/marketink/user/UserService.kt
@@ -1,7 +1,5 @@
 package com.demo.marketink.user
 
-import com.demo.marketink.user.dto.CreateOrUpdateUserDto
-import java.util.Optional
 import java.util.UUID
 import javax.transaction.Transactional
 import org.springframework.stereotype.Service
@@ -11,6 +9,9 @@ class UserService(
     private val userRepository: UserRepository,
 ) {
     fun save(dto: CreateOrUpdateUserDto) {
+        if (findByName(dto.name).isPresent)
+            throw UserNameAlreadyInUseException()
+
         userRepository.save(
             User(
                 externalId = UUID.randomUUID(),
@@ -22,10 +23,17 @@ class UserService(
 
     fun list(): List<User> = userRepository.findAll()
 
-    fun findByExternalId(externalId: UUID): Optional<User> = userRepository.findByExternalId(externalId)
+    fun findByExternalId(externalId: UUID) = userRepository.findByExternalId(externalId)
+
+    fun findByName(name: String) = userRepository.findByName(name);
 
     @Transactional
     fun update(externalId: UUID, dto: CreateOrUpdateUserDto) {
+        if (findByExternalId(externalId).isEmpty)
+            throw UserNotFoundException()
+        if (findByName(dto.name).isPresent)
+            throw UserNameAlreadyInUseException()
+
         findByExternalId(externalId).ifPresent {
             it.name = dto.name
             it.password = dto.password
@@ -34,6 +42,9 @@ class UserService(
 
     @Transactional
     fun delete(externalId: UUID) {
+        if (findByExternalId(externalId).isEmpty)
+            throw UserNotFoundException()
+
         val user = findByExternalId(externalId)
         user.ifPresent {
             it.delete()
